Make `.recv` return `SdkError` instead of `BoxError`
